feat(inline): Add -i, --inline flag for sort command
The inline command (`-i`, `--inline`) allows users to edit the files in
place instead of specifying the output directory.

This command conflicts with the group-by-type and output-dir flags as
the group-by-type will create different files based on the resource
types. The output-dir is implicitly set by the inline flag to the input
directory, so providing the output-dir would be ignored.

Signed-off-by: Dan Thagard <dthagard@gmail.com>